Refactor: Moved the screen code to its own module
